sdk/python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "subtick-sdk"
version = "0.1.0"
description = "Subtick SDK — thin Python wrapper for the Subtick HTTP/WS API (Phase 1)."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "Subtick" }]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests>=2.31",
    "websocket-client>=1.7",
]

[project.urls]
Repository = "https://github.com/subtick/sdk"

[tool.setuptools.packages.find]
include = ["subtick_sdk*"]
exclude = ["examples*", "tests*"]
